refactor(core): Move module logic to `@n8n/backend-common`
