+++ packages/@n8n/backend-common/package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "@n8n/config": "workspace:^",
     "@n8n/constants": "workspace:^",
+    "@n8n/decorators": "workspace:^",
     "@n8n/di": "workspace:^",
     "callsites": "catalog:",
     "n8n-workflow": "workspace:^",
