+++ packages/@n8n/backend-common/src/index.ts
@@ -4,3 +4,5 @@ export * from './types';
 export { inDevelopment, inProduction, inTest } from './environment';
 export { isObjectLiteral } from './utils/is-object-literal';
 export { Logger } from './logging/logger';
+export { ModuleRegistry } from './modules/module-registry';
+export { ModulesConfig, ModuleName } from './modules/modules.config';
