+++ packages/cli/src/execution-lifecycle/execution-lifecycle-hooks.ts
@@ -1,6 +1,7 @@
 import { Logger } from '@n8n/backend-common';
 import { ExecutionRepository } from '@n8n/db';
-import { Container } from '@n8n/di';
+import { LifecycleMetadata } from '@n8n/decorators';
+import { Container, Service } from '@n8n/di';
 import { stringify } from 'flatted';
 import { ErrorReporter, InstanceSettings, ExecutionLifecycleHooks } from 'n8n-core';
 import type {
@@ -11,7 +12,6 @@ import type {
 
 import { EventService } from '@/events/event.service';
 import { ExternalHooks } from '@/external-hooks';
-import { ModuleRegistry } from '@/modules/module-registry';
 import { Push } from '@/push';
 import { WorkflowStatisticsService } from '@/services/workflow-statistics.service';
 import { isWorkflowIdValid } from '@/utils';
@@ -27,6 +27,73 @@ import {
 	updateExistingExecution,
 } from './shared/shared-hook-functions';
 import { type ExecutionSaveSettings, toSaveSettings } from './to-save-settings';
+
+@Service()
+class ModulesHooksRegistry {
+	addHooks(hooks: unknown) {
+		const typedHooks = hooks as ExecutionLifecycleHooks;
+		const handlers = Container.get(LifecycleMetadata).getHandlers();
+
+		for (const { handlerClass, methodName, eventName } of handlers) {
+			const instance = Container.get(handlerClass);
+
+			switch (eventName) {
+				case 'workflowExecuteAfter':
+					typedHooks.addHandler(eventName, async function (runData, newStaticData) {
+						const context = {
+							type: 'workflowExecuteAfter',
+							workflow: this.workflowData,
+							runData,
+							newStaticData,
+						};
+						// eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/return-await
+						return await instance[methodName].call(instance, context);
+					});
+					break;
+
+				case 'nodeExecuteBefore':
+					typedHooks.addHandler(eventName, async function (nodeName, taskData) {
+						const context = {
+							type: 'nodeExecuteBefore',
+							workflow: this.workflowData,
+							nodeName,
+							taskData,
+						};
+						// eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/return-await
+						return await instance[methodName].call(instance, context);
+					});
+					break;
+
+				case 'nodeExecuteAfter':
+					typedHooks.addHandler(eventName, async function (nodeName, taskData, executionData) {
+						const context = {
+							type: 'nodeExecuteAfter',
+							workflow: this.workflowData,
+							nodeName,
+							taskData,
+							executionData,
+						};
+						// eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/return-await
+						return await instance[methodName].call(instance, context);
+					});
+					break;
+
+				case 'workflowExecuteBefore':
+					typedHooks.addHandler(eventName, async function (workflowInstance, executionData) {
+						const context = {
+							type: 'workflowExecuteBefore',
+							workflow: this.workflowData,
+							workflowInstance,
+							executionData,
+						};
+						// eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/return-await
+						return await instance[methodName].call(instance, context);
+					});
+					break;
+			}
+		}
+	}
+}
 
 type HooksSetupParameters = {
 	saveSettings: ExecutionSaveSettings;
@@ -425,7 +492,7 @@ export function getLifecycleHooksForScalingWorker(
 		hookFunctionsPush(hooks, optionalParameters);
 	}
 
-	Container.get(ModuleRegistry).registerLifecycleHooks(hooks);
+	Container.get(ModulesHooksRegistry).addHooks(hooks);
 
 	return hooks;
 }
@@ -487,7 +554,7 @@ export function getLifecycleHooksForScalingMain(
 	hooks.handlers.nodeExecuteBefore = [];
 	hooks.handlers.nodeExecuteAfter = [];
 
-	Container.get(ModuleRegistry).registerLifecycleHooks(hooks);
+	Container.get(ModulesHooksRegistry).addHooks(hooks);
 
 	return hooks;
 }
@@ -511,6 +578,6 @@ export function getLifecycleHooksForRegularMain(
 	hookFunctionsSaveProgress(hooks, optionalParameters);
 	hookFunctionsStatistics(hooks);
 	hookFunctionsExternalHooks(hooks);
-	Container.get(ModuleRegistry).registerLifecycleHooks(hooks);
+	Container.get(ModulesHooksRegistry).addHooks(hooks);
 	return hooks;
 }
