+++ packages/cli/test/integration/shared/test-modules.ts
@@ -1,7 +1,6 @@
+import { ModuleRegistry } from '@n8n/backend-common';
+import type { ModuleName } from '@n8n/backend-common';
 import { Container } from '@n8n/di';
-
-import { ModuleRegistry } from '@/modules/module-registry';
-import type { ModuleName } from '@/modules/modules.config';
 
 export async function loadModules(moduleNames: ModuleName[]) {
 	await Container.get(ModuleRegistry).loadModules(moduleNames);
